Add files via upload
In this project the 'heart' dataset from kaggle was used to predict whether someone would have heart disease based on features such as age. This involved exploratory data analysis to explore the data, creating various visualisations, to gain useful insights about the data. Featuring engineering was then used to decide on the best features to be used for the machine learning algorithms, based on these insights. Before implementing machine learning techniques the data was split into training and testing sets. Four different machine learning algorithms were then used to decide on the best algorithm for this purpose, including the K nearest neighbour. Each algorithm was then evaluated based on their accuracy, precision, recall, f1 score and a confusion matrix, visualised using a heat map. A Kfold and cross validation scoring method was also used to evaluate the algorithms. It was found that the best algorithm for this purpose was the K nearest neighbour algorithm, which was then optimised to find the optimal number of neighbours, achieving a final accuracy of 80%.